upgrade: fix propose replication config script
The script wasn't use a if statement to check for replication.config
file and it was failing when the file exists (since 2.2.1).

+++ config/scripts/propose_replication_config.sh
@@ -4,7 +4,7 @@ set -x
 set -e
 
 # Only execute that if the config repository does not embed replication.config yet
-[ ! -f /root/config/gerrit/replication.config ] && {
+if [ ! -f /root/config/gerrit/replication.config ]; then
     mkdir /root/config/gerrit || true
     cp /home/gerrit/site_path/etc/replication.config /root/config/gerrit/
     cd /root/config
@@ -13,4 +13,4 @@ set -e
         git commit -m "Add gerrit/replication.config in the config repository"
         git review -i
     fi
-}
+fi
